+++ server/src/main/java/com/ones/admin/system/ApiResourceService.java
@@ -276,7 +276,7 @@ public class ApiResourceService {
             SystemMenuMapper menuMapper,
             SystemApiManifestSnapshotMapper manifestSnapshotMapper,
             ObjectMapper objectMapper,
-            @Value("${ones.version:v0.0.37}") String applicationVersion
+            @Value("${ones.version:v0.0.38}") String applicationVersion
     ) {
         this.requestMappingHandlerMapping = requestMappingHandlerMapping;
         this.permissionMapper = permissionMapper;